backend/requirements.txt
# FastAPI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
chromadb>=0.4.22
sqlite-utils>=3.35

# LLM
google-generativeai>=0.3.2

# Data Processing
pandas>=2.1.0
numpy>=1.26.0

# Text Processing
langchain>=0.1.0
langchain-community>=0.0.10
sentence-transformers>=2.2.2

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6

# Testing
pytest>=7.4.0
httpx>=0.26.0
